+++ examples/stackwright-docs/next.config.js
@@ -5,12 +5,6 @@ const withBundleAnalyzer = require('@next/bundle-analyzer')({
 });
 
 module.exports = withBundleAnalyzer(createStackwrightNextConfig({
-    transpilePackages: [
-        "@stackwright/core",
-        "@stackwright/nextjs",
-        "@stackwright/themes",
-        "@stackwright/types",
-    ],
     // Enable static export for R2/CDN hosting
     output: 'export',
     // Images must be unoptimized for static export