+++ packages/hooks-registry/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackwright/hooks-registry
 
+## 0.1.2-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [ed64fab]
+  - @stackwright/types@1.6.0-alpha.3
+
+## 0.1.2-alpha.2
+
+### Patch Changes
+
+- @stackwright/types@1.6.0-alpha.2
+
 ## 0.1.2-alpha.1
 
 ### Patch Changes