+++ packages/types/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @stackwright/types
 
+## 1.6.0-alpha.3
+
+### Minor Changes
+
+- ed64fab: feat(core,types): add format: markdown to TextBlock for CommonMark rendering via micromark
+
+## 1.6.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [a72f3ad]
+  - @stackwright/themes@0.6.0-alpha.0
+
 ## 1.6.0-alpha.1
 
 ### Minor Changes